Create SpellEnums.h
